Refactor to extract call method for multibranch
